requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0

# MLOps
dvc>=3.0.0
mlflow>=2.8.0

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
pydantic>=2.0.0
joblib>=1.3.0
python-multipart>=0.0.6
gunicorn>=21.0.0
requests>=2.31.0

# ML Models
xgboost==1.7.6
optuna>=3.4.0

# Development & Testing
pytest>=7.4.0
flake8>=6.1.0
black>=23.10.0
isort>=5.12.0
